pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "jobctl"
version = "0.1.0"
description = "AI-powered job search assistant CLI and TUI."
readme = "README.md"
requires-python = ">=3.11, <3.14"
dependencies = [
    "click==8.1.8",
    "httpx==0.28.1",
    "jinja2==3.1.6",
    "langgraph==1.1.8",
    "numpy==1.26.4",
    "openai==1.76.2",
    "playwright==1.52.0",
    "pydantic==2.13.1",
    "pygithub==2.6.1",
    "pymupdf==1.26.6",
    "python-docx==1.2.0",
    "pyyaml==6.0.2",
    "qdrant-client==1.12.0",
    "rich==14.0.0",
    "sqlalchemy==2.0.45",
    "alembic==1.17.2",
    "textual==3.1.0",
    "torch==2.2.2",
    "transformers==4.51.3",
    "typer==0.12.5",
    "weasyprint==65.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.5",
    "pytest-asyncio==0.24.0",
    "ruff==0.11.6",
]

[project.scripts]
jobctl = "jobctl.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/jobctl"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
